Force pixel theme by default
While not pretty, this makes the pixel theme to be set on boot as it should
and seen on other devices. The prop itself is making it over but for some reason
is not read in time and so the teal theme is set instead of the pixel one.

Fix 2 indentations aswell

Change-Id: I9c9453a222b53fd5ef821330a34cb9e938e4d9c4